django auth/pyproject.toml
[tool.poetry]
name = "django auth"
version = "0.1.0"
description = ""
authors = ["Nicolas M. <mine.nicolas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-tenants = "^3.7.0"
django-oauth-toolkit = "^3.0.1"
django-allauth = "^65.3.1"
psycopg2 = "^2.9.10"
dj-rest-auth = "^7.0.1"
djangorestframework-simplejwt = "^5.4.0"
mozilla-django-oidc = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
